+++ src/components/layouts/RootLayout.tsx
@@ -2,6 +2,7 @@ import { SearchBar } from "@/components/part/searchbar"
 import SimpleSidebar from "@/components/part/sidebar"
 import { chakraCustomTheme } from "@/themes/theme"
 import { ChakraProvider } from "@chakra-ui/react"
+
 type LayoutProps = {
 	children: React.ReactNode
 }
