+++ src/components/part/landing.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/no-unescaped-entities */
 import { useState } from "react"
-import styles from "../../styles/landing-style.module.css"
+import styles from "@/styles/landing-style.module.css"
 
 type CardProps = {
 	title: string
