+++ src/components/part/mill/Icons.tsx
@@ -1,4 +1,3 @@
-
 interface IconProps {
 	className?: string
 }
