+++ src/components/part/trending/trendingArticleBox.tsx
@@ -6,7 +6,7 @@ import { useTrendingNote, useTrendingNoteSlide } from "@/state/TrendingNote"
 import { Box, Skeleton, SkeletonCircle, SkeletonText } from "@chakra-ui/react"
 import { CharacterAvatar } from "@crossbell/ui"
 import ChakraUIRenderer from "chakra-ui-markdown-renderer"
-import { Note } from "crossbell"
+import type { Note } from "crossbell"
 import { useEffect, useRef, useState } from "react"
 import { FiArrowUp } from "react-icons/fi"
 import ReactMarkdown from "react-markdown"
